requirements.txt
pytest>=7.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
black>=23.0.0